+++ RandomSkunk.Results.UnitTests/FailIf_methods.cs
@@ -1,0 +1,129 @@
+namespace RandomSkunk.Results.UnitTests;
+
+public class FailIf_methods
+{
+    public class For_Result
+    {
+        [Fact]
+        public void GivenSuccessResult_WhenPredicateReturnsTrue_ReturnsFailResult()
+        {
+            var result = Result.Success();
+            var error = new Error();
+
+            var actual = result.FailIf(() => true, () => error);
+
+            actual.IsFail.Should().BeTrue();
+            actual.Error.Message.Should().BeSameAs(error.Message);
+            actual.Error.Title.Should().BeSameAs(error.Title);
+        }
+
+        [Fact]
+        public void GivenSuccessResult_WhenPredicateReturnsFalse_ReturnsSameResult()
+        {
+            var result = Result.Success();
+            var error = new Error();
+
+            var actual = result.FailIf(() => false, () => error);
+
+            actual.IsSuccess.Should().BeTrue();
+        }
+
+        [Fact]
+        public void GivenFailResult_WhenPredicateReturnsTrue_ReturnsSameResult()
+        {
+            var result = Result.Fail();
+            var error = new Error();
+
+            var actual = result.FailIf(() => true, () => error);
+
+            actual.Should().Be(result);
+        }
+    }
+
+    public class For_Result_of_T
+    {
+        [Fact]
+        public void GivenSuccessResult_WhenPredicateReturnsTrue_ReturnsFailResult()
+        {
+            var result = Result<int>.Success(123);
+            var error = new Error();
+
+            var actual = result.FailIf(value => true, value => error);
+
+            actual.IsFail.Should().BeTrue();
+            actual.Error.Message.Should().BeSameAs(error.Message);
+            actual.Error.Title.Should().BeSameAs(error.Title);
+        }
+
+        [Fact]
+        public void GivenSuccessResult_WhenPredicateReturnsFalse_ReturnsSameResult()
+        {
+            var result = Result<int>.Success(123);
+            var error = new Error();
+
+            var actual = result.FailIf(value => false, value => error);
+
+            actual.IsSuccess.Should().BeTrue();
+        }
+
+        [Fact]
+        public void GivenFailResult_WhenPredicateReturnsTrue_ReturnsSameResult()
+        {
+            var result = Result<int>.Fail();
+            var error = new Error();
+
+            var actual = result.FailIf(value => true, value => error);
+
+            actual.Should().Be(result);
+        }
+    }
+
+    public class For_Maybe_of_T
+    {
+        [Fact]
+        public void GivenSuccessResult_WhenPredicateReturnsTrue_ReturnsFailResult()
+        {
+            var result = Maybe<int>.Success(123);
+            var error = new Error();
+
+            var actual = result.FailIf(value => true, value => error);
+
+            actual.IsFail.Should().BeTrue();
+            actual.Error.Message.Should().BeSameAs(error.Message);
+            actual.Error.Title.Should().BeSameAs(error.Title);
+        }
+
+        [Fact]
+        public void GivenSuccessResult_WhenPredicateReturnsFalse_ReturnsSameResult()
+        {
+            var result = Maybe<int>.Success(123);
+            var error = new Error();
+
+            var actual = result.FailIf(value => false, value => error);
+
+            actual.IsSuccess.Should().BeTrue();
+        }
+
+        [Fact]
+        public void GivenFailResult_WhenPredicateReturnsTrue_ReturnsSameResult()
+        {
+            var result = Maybe<int>.Fail();
+            var error = new Error();
+
+            var actual = result.FailIf(value => true, value => error);
+
+            actual.Should().Be(result);
+        }
+
+        [Fact]
+        public void GivenNoneResult_WhenPredicateReturnsTrue_ReturnsSameResult()
+        {
+            var result = Maybe<int>.None();
+            var error = new Error();
+
+            var actual = result.FailIf(value => true, value => error);
+
+            actual.Should().Be(result);
+        }
+    }
+}