+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning].
 - Add support for .NET 7.
 - RandomSkunk.Results
     - Add `[TryCatch]` and `[assembly: TryCatchThirdParty]` attributes.
+    - Add `FailIf` methods.
 - RandomSkunk.Results.Analyzers (new package)
     - Add try/catch source generator.